requirements.txt
anyio==4.4.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.9
markdownify==0.13.1
ollama==0.3.3
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.2.3
